Move Project to own entity namespace
